+++ etrm-system/etrm-backend/src/main/java/com/etrm/system/rbac/AppModule.java
@@ -20,7 +20,7 @@ public class AppModule {
     private String description;
 
     @Column(name = "sort_order", nullable = false)
-    private Integer sortOrder = 0;
+    private Short sortOrder = 0;
 
     @Column(name = "is_active", nullable = false)
     private Boolean isActive = true;
@@ -33,8 +33,8 @@ public class AppModule {
     public void setModuleName(String moduleName) { this.moduleName = moduleName; }
     public String getDescription() { return description; }
     public void setDescription(String description) { this.description = description; }
-    public Integer getSortOrder() { return sortOrder; }
-    public void setSortOrder(Integer sortOrder) { this.sortOrder = sortOrder; }
+    public Short getSortOrder() { return sortOrder; }
+    public void setSortOrder(Short sortOrder) { this.sortOrder = sortOrder; }
     public Boolean getIsActive() { return isActive; }
     public void setIsActive(Boolean isActive) { this.isActive = isActive; }
 }
